SetBuilders

Documentation for SetBuilders.

SetBuilders.@setconvert

SetBuilders.@setenum

SetBuilders.@setfilter

SetBuilders.@setconvert — Macro

A set converter from a set to another

struct MyStruct
    a
end

I = SB_SET_INT
S = setfromtype(MyStruct)

A = @setconvert(y in S, x -> mystruct(x), y -> y.a, x in I,
                mystruct=MyStruct)

is similar to the mathematical set notation of 

A = {y ∈ S | y = MyStruct(x) and x = y.a where x ∈ I}

@assert MyStruct(1) in A
@assert !(MyStruct(1.0) in A)

, and where I is the integer set and S is the set of all possible
MyStruct objects. SB_SET_INT is a pre-defined set in SetBuilders.jl
package for all Julia integer data types such as Int64 and Int32.

SetBuilders.@setenum — Macro

A enumerated set builder

A = @setenum([1, 2, 3])

is similar to the mathematical set notation of 

A = {1, 2, 3}

@assert 1 in A
@assert !(0 in A)

This EnumSet wraps Julia Set objects.

SetBuilders.@setfilter — Macro

A filtered set builder

I = SB_SET_INT

A = @setfilter((x in I, y in I), x > 0 && y > 0)

is similar to the mathematical set notation of 

A = {x ∈ I, y ∈ I | x > 0 and y > 0}

@assert (1, 1) in A
@assert !((0, 0) in A)

, where I is the integer set. SB_SET_INT is a pre-defined set
in SetBuilders.jl package for all Julia integer data types
such as Int64 and Int32.