SetBuilders

Documentation for SetBuilders.

SetBuilders.@setfilter

SetBuilders.@setmap

SetBuilders.@setfilter — Macro

A filtered set builder

A = @setfilter((x in SB_SET_INT64, y in SB_SET_INT64), x > 0 && y > 0)

is similar to the mathematical set definiton of 

A = {x ∈ I, y ∈ I | x > 0 and y > 0}

, where I is the integer set. SB_SET_INT64 is a pre-defined set
in SetBuilders.jl package for Julia Int64 data type.

SetBuilders.@setmap — Macro

A mapped set builder struct MyStruct a end

A = @setmap(x in SB_SET_INT64, x -> my_struct(x), MyStruct,
            my_struct=MyStrict)

A = @setmap(y in MyStruct, x -> my_struct(x), x=SB_SET_INT64, my_struct=MyStrict)

is similar to the mathematical set definiton of 

A = {y ∈ MyStruct | y = MyStruct(x) where x ∈ I}

, where I is the integer set and MyStruct is the all possible
MyStruct objects. SB_SET_INT64 is a pre-defined set in SetBuilders.jl
package for Julia Int64 data type.