Reference

SetBuilders.SetBuilders — Module

Main module for SetBuilders – Julia Package for Predicate and Enumerable Sets

Exports

@setbuild : builds a SetBuilders set

@setpkg : loads a set from a Julia module

ismember/in/∈ : checks if an object is a member of a set

describe : generates a set description string

fmap : generates an element in codomain of a MappedSet

bmap : generates an element in domain of a MappedSet

union/∪ : set union operation

intersection/∩: set union operation

setdiff/- : set difference operation

symdiff : set symmetric difference operation

complement/~ : set complement operation

push! : adds an element to an EnumerableSet

pop! : removes an element from an EnumerableSet

SBSet : the top level type of all SetBuilders sets

SetBuilders.SBSet — Type

SBSet - Type

The SBSet type is the supertype of all SetBuilders set types.

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> I isa SBSet
true

SetBuilders.ismember — Method

ismember(elem, set::CompositeSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> A = @setbuild(x in I, 0 <= x < 10)
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)

julia> B = @setbuild(x in I, 5 <= x < 15)
PredicateSet((x ∈ TypeSet(Integer)) where 5 <= x < 15)

julia> C = A ∩ B
CompositeSet(PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10) ∩ PredicateSet((x ∈ TypeSet(Integer)) where 5 <= x < 15))

julia> ismember(5, C) # 5 in C
true

julia> ismember(0, C) # 0 in C
false

SetBuilders.ismember — Method

ismember(elem, set::EmptySet; kwargs...)

Check if elem is a member of set

Examples

julia> E = @setbuild()
EmptySet()

julia> ismember(1, E)   # 1 in E
false

julia> ismember(0.1, E) # 0.1 in E
false

SetBuilders.ismember — Method

ismember(elem, set::EnumerableSet; kwargs...)

Check if elem is a member of set

Examples

julia> A = @setbuild(Union{Int64, Float64}[1])
EnumerableSet([{Float64}*0, {Int64}*1])

julia> ismember(1, A)
true

julia> ismember(Int32(1), A)
false

julia> push!(A, Float64(2.0))
EnumerableSet([{Float64}*1, {Int64}*1])

julia> ismember(Float64(2.0), A)
true

julia> pop!(A, Float64(2.0))
2.0

julia> ismember(Float64(2.0), A)
false

SetBuilders.ismember — Method

ismember(elem, set::MappedSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> struct MyStruct
       a
       b
       end

julia> S = @setbuild(MyStruct)
TypeSet(MyStruct)

julia> A = @setbuild(s in S, (x in I, y in I) -> mystruct(x,y), s -> (s.a, s.b),
                     mystruct=MyStruct)
MappedSet((x ∈ TypeSet(Integer)), (y ∈ TypeSet(Integer)) -> (s ∈ TypeSet(MyStruct)))

julia> ismember(MyStruct(1, 1), A)   # MyStruct(1, 1) in A
true

julia> ismember(MyStruct(1.0, 1), A) # MyStruct(1.0, 1) in A
false

SetBuilders.ismember — Method

ismember(elem, set::PredicateSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> A = @setbuild(x in I, 0 <= x < 10)
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)

julia> ismember(0, A)  # 0 in A 
true

julia> ismember(10, A) # 10 in A
false

SetBuilders.ismember — Method

ismember(elem, set::TypeSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> ismember(1, I)   # 1 in I
true

julia> ismember(0.1, I) # 0.1 in I
false

SetBuilders.ismember — Method

ismember(elem, set::UniversalSet; kwargs...)

Check if elem is a member of set

Examples

julia> U = @setbuild(Any)
UniversalSet()

julia> ismember(1, U)   # 1 in U
true

julia> ismember(0.1, U) # 0.1 in U
true

SetBuilders.@setbuild — Macro

@setbuild([args...[; kwargs...]])

The @setbuild macro creates various SetBuilders sets.

The number of arguments varies depending on the type of set to create.

No argument

EmptySet

julia> @setbuild()
EmptySet()

One argument

TypeSet

The first argument is any Julia type.

julia> @setbuild(Integer)
TypeSet(Integer)

EnumerableSet

The first argument is a Vector with optionally Julia type

julia> @setbuild([1, 2, 3])
EnumerableSet([{Int64}*3])

julia> @setbuild(Int32[])
EnumerableSet([{Int32}*0])

UniversalSet

The first argument is Any type

julia> @setbuild(Any)
UniversalSet()

Two arguments

PredicateSet

The first argument is the domain of the set. The second argument is the predicate of the set.

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> @setbuild(x in I, 0 <= x < 5)
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 5)

Four arguments

MappedSet

The first argument is the domain of the set. The second argument is the codomain of the set. The third argument is the forward mapping of the set. The fourth argument is the backward mapping of the set.

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> @setbuild(x in I, y in I, y = x + 1, x = y - 1)
MappedSet((x ∈ TypeSet(Integer)) -> (y ∈ TypeSet(Integer)))

Keyword arguments

The main usage of keyword arguments is to provide objects defined outside of @setbuild to expressions inside of @setbuild.

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> k = 1
1

julia> @setbuild(x in I, y in I, y = x + c, x = y - c, c=k)
MappedSet((x ∈ TypeSet(Integer)) -> (y ∈ TypeSet(Integer)))

Note

keyward names that start with "sb_" are reserved names for SetBuilders.

SetBuilders.@setpkg — Macro

@setpkg command[ command-arguments... ]

The @setpkg macro enables the reuse of sets that were developed separately.

commands

load: loads sets from a local file, also known as a setfile

@setpkg load <path/to/file>

The setfile is a regular Julia module customized for SetBuilders.

Examples

Assuming that the file myset.sjl contains the following Julia code:

module MySetModule

export MYSET

I = @setbuild(Integer)
MYSET = @setbuild(x in I, x > 0)

end

MYSET can be used as shown in the example below:

julia> @setpkg load "myset.sjl"

julia> using SetBuilders.MySetModule

julia> 1 in MYSET
true

julia> 0 in MYSET
false