Reference

SetBuilders.SetBuilders — Module

Main module for SetBuilders.jl – predicate-based set generation package for Julia.

SetBuilders.SBSet — Type

SBSet - Type

The SBSet type is the supertype of all SetBuilders set types.

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> I isa SBSet
true

SetBuilders.ismember — Method

ismember(elem, set::CompositeSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> A = @setbuild(x in I, 0 <= x < 10)
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)

julia> B = @setbuild(x in I, 5 <= x < 15)
PredicateSet((x ∈ TypeSet(Integer)) where 5 <= x < 15)

julia> C = A ∩ B
CompositeSet(PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10) ∩ PredicateSet((x ∈ TypeSet(Integer)) where 5 <= x < 15))

julia> ismember(5, C) # 5 in C
true

julia> ismember(0, C) # 0 in C
false

SetBuilders.ismember — Method

ismember(elem, set::EmptySet; kwargs...)

Check if elem is a member of set

Examples

julia> E = @setbuild()
EmptySet()

julia> ismember(1, E)   # 1 in E
false

julia> ismember(0.1, E) # 0.1 in E
false

SetBuilders.ismember — Method

ismember(elem, set::EnumerableSet; kwargs...)

Check if elem is a member of set

Examples

julia> A = @setbuild(Union{Int64, Float64}[1])
EnumerableSet([{Float64}*0, {Int64}*1])

julia> ismember(1, A)
true

julia> ismember(Int32(1), A)
false

julia> push!(A, Float64(2.0))
EnumerableSet([{Float64}*1, {Int64}*1])

julia> ismember(Float64(2.0), A)
true

julia> pop!(A, Float64(2.0))
2.0

julia> ismember(Float64(2.0), A)
false

SetBuilders.ismember — Method

ismember(elem, set::MappedSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> struct MyStruct
       a
       b
       end

julia> S = @setbuild(MyStruct)
TypeSet(MyStruct)

julia> A = @setbuild(s in S, (x in I, y in I) -> mystruct(x,y), s -> (s.a, s.b),
                     mystruct=MyStruct)
MappedSet((x ∈ TypeSet(Integer)), (y ∈ TypeSet(Integer)) -> (s ∈ TypeSet(MyStruct)))

julia> ismember(MyStruct(1, 1), A)   # MyStruct(1, 1) in A
true

julia> ismember(MyStruct(1.0, 1), A) # MyStruct(1.0, 1) in A
false

SetBuilders.ismember — Method

ismember(elem, set::PredicateSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> A = @setbuild(x in I, 0 <= x < 10)
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)

julia> ismember(0, A)  # 0 in A 
true

julia> ismember(10, A) # 10 in A
false

SetBuilders.ismember — Method

ismember(elem, set::TypeSet; kwargs...)

Check if elem is a member of set

Examples

julia> I = @setbuild(Integer)
TypeSet(Integer)

julia> ismember(1, I)   # 1 in I
true

julia> ismember(0.1, I) # 0.1 in I
false

SetBuilders.ismember — Method

ismember(elem, set::UniversalSet; kwargs...)

Check if elem is a member of set

Examples

julia> U = @setbuild(Any)
UniversalSet()

julia> ismember(1, U)   # 1 in U
true

julia> ismember(0.1, U) # 0.1 in U
true

SetBuilders.@setbuild — Macro

@setbuild([args...[; kwargs...]])

The @setbuild macro creates various SetBuilders sets.

The @setbuild macro in SetBuilders for creating sets from Julia data types, predicates, and mappings. For example, I = @setbuild(Integer) creates a set of all Julia Integer type objects, and A = @setbuild(x ∈ I, 0 < x < 4) creates a set that implies to contain the integers 1, 2, and 3.

Examples

julia> E = @setbuild()
EmptySet()

julia> U = @setbuild(Any)
UniversalSet()

julia> I = @setbuild(Integer) # Julia Integer-type set
TypeSet(Integer)

julia> D = @setbuild(Dict{String, Number}) # Julia Dict{String, Number}-type set
TypeSet(Dict{String, Number})

julia> struct MyStruct
           a
           b
       end

julia> S = @setbuild(MyStruct)  # Julia user-type set
TypeSet(MyStruct)

julia> N = @setbuild([1, 2, 3]) # Enumerable set
EnumerableSet([{Int64}*3])

julia> C = @setbuild((I, I))  # Cartesian sets
PredicateSet((c1 ∈ TypeSet(Integer)), (c2 ∈ TypeSet(Integer)) where true)

julia> P = @setbuild(x in I, 0 <= x < 10) # Predicate sets
PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)

julia> M = @setbuild(z in I, (x in P) -> x + 5, z -> z - 5) # Mapped sets
MappedSet((x ∈ PredicateSet((x ∈ TypeSet(Integer)) where 0 <= x < 10)) -> (z ∈ TypeSet(Integer)))

SetBuilders.@setpkg — Macro

@setpkg command[ command-arguments... ]

The @setpkg macro enables the reuse of sets that were developed separately.

commands

load: loads sets from a local file, also known as a setfile

@setpkg load <path/to/file>

The setfile is a regular Julia module customized for SetBuilders.

Examples

Assuming that the file myset.sjl contains the following Julia code:

module MySetModule

export MYSET

I = @setbuild(Integer)
MYSET = @setbuild(x in I, x > 0)

end

MYSET can be used as shown in the example below:

julia> @setpkg load "myset.sjl"

julia> using SetBuilders.MySetModule

julia> 1 in MYSET
true

julia> 0 in MYSET
false